% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpcop-package.R
\docType{package}
\name{Rpcop-package}
\alias{Rpcop}
\alias{Rpcop-package}
\title{Principal Curves of Oriented Points}
\description{
Core package documentation generated by roxygen2.
}
\author{
\strong{Maintainer}: Kevin Michael Frick \email{kmfrick@proton.me} (Modern C++ fixes and Rcpp port, with permission from the original authors) [translator]

Authors:
\itemize{
  \item Pedro Delicado \email{pedro.delicado@upc.edu} (Original C++ author)
  \item Mario Huerta \email{mario.huerta@uab.es} (Original C++ author)
}

Other contributors:
\itemize{
  \item Stephen L. Moshier \email{steve@moshier.net} (Author of eigens() for symmetric matrix eigendecomposition) [copyright holder]
}

}
\keyword{internal}
