% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpcop.R
\name{pcop}
\alias{pcop}
\title{Principal Curve of Oriented Points}
\usage{
pcop(x, Ch = 1.5, Cd = 0.3, plot.true = FALSE, ...)
}
\arguments{
\item{x}{A numeric matrix of \eqn{n} points in dimension \eqn{p}.}

\item{Ch}{The smoothing parameter \eqn{h} is \eqn{C_H} times the value given by
the normal reference rule. Default value \eqn{1.5}. Constraints
\eqn{0.5 \le C_H \le 1.5}.}

\item{Cd}{The distance between two consecutive principal oriented points in a
PCOP is about \eqn{C_D} times the value of the smoothing parameter
\eqn{h}. Default value \eqn{0.3}. Constraints \eqn{0.25 \le C_D \le 0.5}.}

\item{plot.true}{If \code{TRUE}, produce a plot of the resulting curve.}

\item{...}{Additional parameters passed to \code{lines}.}
}
\value{
A list with two elements:
\describe{
  \item{pcop.f1}{Data frame storing the principal curve of oriented points in
    the original format, with columns \code{param}, \code{dens},
    \code{span}, \code{orth.var}, \code{pop1}, \code{pop2}, \ldots,
    \code{pr.dir1}, \code{pr.dir2}, \ldots}
  \item{pcop.f2}{List conforming to the format used in \pkg{princurve}; see that
    package for details.}
}
}
\description{
Computes a principal curve as defined in Delicado (2001)
\doi{10.1007/s001800300145}.
}
\examples{
n <- 500
p <- 3
x <- matrix(rnorm(n * p), ncol = p) \%*\% diag(p:1)
pcop(x, plot.true = TRUE, lwd = 4, col = 2)

x <- runif(100, -1, 1)
x <- cbind(x, x ^ 2 + rnorm(100, sd = 0.1))
pcop(x, plot.true = TRUE, lwd = 4, col = 2)
}
