% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcop_backend}
\alias{pcop_backend}
\title{Computes a principal curve as defined in Delicado (2001). DO NOT use this function unless you know what you are doing. Use `pcop()` instead.}
\usage{
pcop_backend(x, c_d, c_h)
}
\arguments{
\item{x}{See `pcop()`}

\item{c_d}{See `pcop()`}

\item{c_h}{See `pcop()`}
}
\value{
A numeric matrix to be parsed by `pcop()`.
}
\description{
Computes a principal curve as defined in Delicado (2001). DO NOT use this function unless you know what you are doing. Use `pcop()` instead.
}
